projects/instagram-automation/requirements.txt
# Instagram Automation System - Python Dependencies

# Core
python-dotenv>=1.0.0
requests>=2.31.0

# Image Processing
Pillow>=10.0.0

# Cloud Storage
boto3>=1.34.0  # AWS S3 / Cloudflare R2

# Optional (for advanced features)
# opencv-python>=4.8.0  # Video frame extraction
# numpy>=1.24.0         # Array operations
